+++ berserk/clients/bots.py
@@ -5,7 +5,7 @@ from typing import Iterator, Any, Dict
 from .. import models
 from ..formats import NDJSON
 from .base import BaseClient
-from ..types.common import ChallengeDeclineReason
+from ..types.challenges import ChallengeDeclineReason
 
 
 class Bots(BaseClient):
