+++ berserk/types/__init__.py
@@ -3,11 +3,10 @@ from __future__ import annotations
 from .account import AccountInformation, Perf, Preferences, Profile, StreamerInfo
 from .bulk_pairings import BulkPairing, BulkPairingGame
 from .challenges import Challenge
-from .common import ClockConfig, ExternalEngine, LightUser, OnlineLightUser
+from .common import ClockConfig, ExternalEngine, LightUser, OnlineLightUser, Variant
 from .puzzles import PuzzleRace
 from .opening_explorer import (
     OpeningExplorerRating,
-    OpeningExplorerVariant,
     OpeningStatistic,
     Speed,
 )
@@ -26,7 +25,7 @@ __all__ = [
     "OnlineLightUser",
     "CurrentTournaments",
     "OpeningExplorerRating",
-    "OpeningExplorerVariant",
+    "Variant",
     "OpeningStatistic",
     "PaginatedTeams",
     "Perf",
