+++ berserk/types/common.py
@@ -47,6 +47,11 @@ GameType: TypeAlias = Literal[
     "fromPosition",
 ]
 
+Speed = Literal[
+    "ultraBullet", "bullet", "blitz", "rapid", "classical", "correspondence"
+]
+
+
 Title = Literal[
     "GM", "WGM", "IM", "WIM", "FM", "WFM", "NM", "CM", "WCM", "WNM", "LM", "BOT"
 ]
@@ -77,27 +82,3 @@ PerfType: TypeAlias = Union[
 GameRule: TypeAlias = Literal[
     "noAbort", "noRematch", "noGiveTime", "noClaimWin", "noEarlyDraw"
 ]
-
-ChallengeDeclineReason: TypeAlias = Literal[
-    "generic",
-    "later",
-    "tooFast",
-    "tooSlow",
-    "timeControl",
-    "rated",
-    "casual",
-    "standard",
-    "variant",
-    "noBot",
-    "onlyBot",
-]
-
-ChallengeStatus: TypeAlias = Literal[
-    "created",
-    "offline",
-    "canceled",
-    "declined",
-    "accepted",
-]
-
-ChallengeDirection: TypeAlias = Literal["in", "out"]
